requirements.txt
# Core
httpx>=0.27
tenacity>=8.0
pyyaml>=6.0
python-dotenv>=1.0

# Progress / display
rich>=13.0

# Analysis & export
pandas>=2.0

# Optional — for parquet export
# pyarrow>=15.0
